Remove mixins
