--- addon/components/adorable.js
@@ -1,4 +1,0 @@
-import Component from '@ember/component';
-import Adorable from 'ember-initials/mixins/adorable';
-
-export default Component.extend(Adorable);
