--- addon/components/gravatar.js
@@ -1,4 +1,0 @@
-import Component from '@ember/component';
-import Gravatar from 'ember-initials/mixins/gravatar';
-
-export default Component.extend(Gravatar);
