--- addon/components/image.js
@@ -1,4 +1,0 @@
-import Component from '@ember/component';
-import Image from 'ember-initials/mixins/image';
-
-export default Component.extend(Image);
