+++ addon/components/image/template.hbs
@@ -1,0 +1,9 @@
+<img ...attributes
+  class={{this.class}}
+  width={{this.width}}
+  height={{this.height}}
+  src={{this.src}}
+  alt={{this.alt}}
+  title={{this.title}}
+  onError={{this.onError}}
+/>
