--- addon/components/initials.js
@@ -1,4 +1,0 @@
-import Component from '@ember/component';
-import Initials from 'ember-initials/mixins/initials';
-
-export default Component.extend(Initials);
