+++ app/components/ember-initials/gravatar/component.js
@@ -1,1 +1,2 @@
-export { default } from 'ember-initials/components/gravatar';
+export { default } from 'ember-initials/components/gravatar/component';
+
