+++ package.json
@@ -34,7 +34,8 @@
   "dependencies": {
     "blueimp-md5": "2.10.0",
     "ember-auto-import": "^1.3.0",
-    "ember-cli-babel": "^7.7.3"
+    "ember-cli-babel": "^7.7.3",
+    "ember-cli-htmlbars": "^3.0.1"
   },
   "devDependencies": {
     "@ember/optional-features": "^0.7.0",
@@ -49,7 +50,6 @@
     "ember-cli-fastboot": "2.0.4",
     "ember-cli-favicon": "2.0.0",
     "ember-cli-github-pages": "0.2.1",
-    "ember-cli-htmlbars": "^3.0.1",
     "ember-cli-htmlbars-inline-precompile": "^2.1.0",
     "ember-cli-inject-live-reload": "^2.0.1",
     "ember-cli-release": "1.0.0-beta.2",
